Fix: Available Page 컴포넌트화#90
